Refactoring and Azure Blob connectivity
