+++ src/ncea-harvester.tests/Clients/KeyVaultServiceForTests.cs
@@ -1,23 +1,22 @@
 ﻿using Azure.Security.KeyVault.Secrets;
 using Azure;
 using Moq;
-using ncea.harvester.infra;
+using Ncea.Harvester.Infrastructure;
 
-namespace ncea_harvester.tests.Clients
+namespace Ncea.Harvester.Tests.Clients;
+
+public static class KeyVaultServiceForTests
 {
-    public static class KeyVaultServiceForTests
+    public static KeyVaultService Get(string key, string value)
     {
-        public static KeyVaultService Get(string key, string value)
-        {
-            var mockSecretClient = new Mock<SecretClient>();
-            var mockSecret = new KeyVaultSecret(key, value);
-            var mockResponse = Response.FromValue(mockSecret, new Mock<Response>().Object);
-            mockSecretClient.Setup<Task<Azure.Response<KeyVaultSecret>>>(x => x.GetSecretAsync(It.IsAny<string>(), null, default))
-                            .ReturnsAsync(mockResponse);
-            var secretClient = mockSecretClient.Object;
-            var clientService = new KeyVaultService(secretClient);
+        var mockSecretClient = new Mock<SecretClient>();
+        var mockSecret = new KeyVaultSecret(key, value);
+        var mockResponse = Response.FromValue(mockSecret, new Mock<Response>().Object);
+        mockSecretClient.Setup<Task<Azure.Response<KeyVaultSecret>>>(x => x.GetSecretAsync(It.IsAny<string>(), null, default))
+                        .ReturnsAsync(mockResponse);
+        var secretClient = mockSecretClient.Object;
+        var clientService = new KeyVaultService(secretClient);
 
-            return clientService;
-        }
+        return clientService;
     }
 }
