+++ src/ncea-harvester.tests/Clients/ServiceBusServiceForTests.cs
@@ -1,24 +1,23 @@
 ﻿using Moq;
 using Azure.Messaging.ServiceBus;
 using Microsoft.Extensions.Options;
-using ncea.harvester.infra;
-using ncea.harvester.Models;
+using Ncea.Harvester.Infrastructure;
+using Ncea.Harvester.Models;
 
-namespace ncea_harvester.tests.Clients
+namespace Ncea.Harvester.Tests.Clients;
+
+public static class ServiceBusServiceForTests
 {
-    public static class ServiceBusServiceForTests
+    public static ServiceBusService Get(out Mock<ServiceBusSender> mockServiceBusSender)
     {
-        public static ServiceBusService Get(out Mock<ServiceBusSender> mockServiceBusSender)
-        {
-            var appSettings = Options.Create(new AppSettings() { ServiceBusConnectionString = "azure.servicebus-connection", ServiceBusQueueName = "test-queue" });
-            var mockServiceBusClient = new Mock<ServiceBusClient>();
-            mockServiceBusSender = new Mock<ServiceBusSender>();
-            var service = new ServiceBusService(appSettings, mockServiceBusClient.Object);
+        var appSettings = Options.Create(new HarvesterConfigurations() { ServiceBusConnectionString = "azure.servicebus-connection", ServiceBusQueueName = "test-queue" });
+        var mockServiceBusClient = new Mock<ServiceBusClient>();
+        mockServiceBusSender = new Mock<ServiceBusSender>();
+        var service = new ServiceBusService(appSettings, mockServiceBusClient.Object);
 
-            // Set up the mock to return the mock sender
-            mockServiceBusClient.Setup(x => x.CreateSender(It.IsAny<string>())).Returns(mockServiceBusSender.Object);
+        // Set up the mock to return the mock sender
+        mockServiceBusClient.Setup(x => x.CreateSender(It.IsAny<string>())).Returns(mockServiceBusSender.Object);
 
-            return service;
-        }
+        return service;
     }
 }
