--- src/ncea-harvester.tests/Infra/KeyVaultServiceTests.cs
@@ -1,20 +1,0 @@
-﻿using ncea_harvester.tests.Clients;
-
-namespace ncea_harvester.tests.Infra
-{
-    public class KeyVaultServiceTests
-    {
-        [Fact]
-        public async Task GetSecretAsync_ShouldReturnSecretValue()
-        {
-            // Arrange
-            var keyVaultService = KeyVaultServiceForTests.Get("test-secret-key", "test-secret-value");
-
-            // Act
-            var result = await keyVaultService.GetSecretAsync("test-secret-key");
-
-            // Assert
-            Assert.Equal("test-secret-value", result);
-        }        
-    }
-}
