+++ src/ncea-harvester/Infrastructure/Contracts/IApiClient.cs
@@ -1,0 +1,8 @@
+﻿namespace Ncea.Harvester.Infrastructure.Contracts;
+
+public interface IApiClient
+{
+    void CreateClient(string BaseUrl);
+    Task<string> GetAsync(string apiUrl);
+    Task<string> PostAsync(string apiUrl, string requestData);
+}
