+++ src/ncea-harvester/Infrastructure/Contracts/IKeyVaultService.cs
@@ -1,0 +1,6 @@
+﻿namespace Ncea.Harvester.Infrastructure.Contracts;
+
+public interface IKeyVaultService
+{
+    Task<string> GetSecretAsync(string key);
+}
