+++ src/ncea-harvester/Infrastructure/Contracts/IServiceBusService.cs
@@ -1,0 +1,6 @@
+﻿namespace Ncea.Harvester.Infrastructure.Contracts;
+
+public interface IServiceBusService
+{
+    Task SendMessageAsync(string message);
+}
