+++ src/ncea-harvester/Infrastructure/Models/Requests/DeleteBlobRequest.cs
@@ -1,0 +1,8 @@
+﻿namespace Ncea.Harvester.Infrastructure.Models.Requests;
+
+public class DeleteBlobRequest
+{
+    public string Blob { get; set; }
+   
+    public string Container { get; set; }
+}
