+++ src/ncea-harvester/Processor/Contracts/IProcessor.cs
@@ -1,0 +1,6 @@
+﻿namespace Ncea.Harvester.Processors.Contracts;
+
+public interface IProcessor
+{
+    Task Process();
+}
