--- src/ncea-harvester/Processors/IProcessor.cs
@@ -1,7 +1,0 @@
-﻿namespace ncea.harvester.Processors
-{
-    public interface IProcessor
-    {
-        Task Process();
-    }
-}
