+++ src/ncea-harvester/appsettings.json
@@ -7,9 +7,10 @@
       "DataSourceApiUrl": "/waf/index.html",
       "Schedule": "*/5 * * * *"
     },
-    "ServiceBusConnectionString": "nceapocsbox.servicebus.windows.net",
+    "ServiceBusHostName": "harvestersb.servicebus.windows.net",
     "ServiceBusQueueName": "jncc-harvester-queue",
-    "KeyVaultUri": "",
+    "KeyVaultUri": "https://nceakv.vault.azure.net",
+    "BlobStorage": "https://nceaharvesterblob.blob.core.windows.net",
     "ApplicationInsights": {
       "ConnectionString": "InstrumentationKey=00000000-0000-0000-0000-000000000000;"
     },
